+++ src/components/JardinConnectes/index.js
@@ -1,12 +1,16 @@
 import React from 'react';
 import Header from 'src/components/Header';
+import Navigation from 'src/components/Navigation';
 
 import './style.scss';
 
 const JardinConnectes = () => {
   console.log('App launched');
   return (
-    <Header />
+    <div className="jardinconnectes">
+      <Header />
+      <Navigation />
+    </div>
   );
 };